Source code of imgconvert
This python script allow users to convert between various standard image formats such as jpg, png, webp and many more. 
After executing this python shell script, a Path is asked for media to convert. If provided path is a file, it asks for output format and does the task of conversion, however if path is a directory, the program asks the final output format of media to be converted and input format type for selection of all the images of same input type provided by the user.